% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jy_text_analyser_main.R
\name{jy_text_analyser}
\alias{jy_text_analyser}
\title{jy_text_analyser (main function)}
\usage{
jy_text_analyser(file_path)
}
\arguments{
\item{file_path}{the file path to text file to be read (file-format must be "*.txt")}
}
\value{
Ratio between positive and negative words (numeric scalar)
}
\description{
A (wrapper) function to read words from a text file, and find the ratio between positive and negative words
}
\examples{
jy_text_analyser("../data/Example_negative.txt")
}
