% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jy_text_file_reader.R
\name{jy_text_file_reader}
\alias{jy_text_file_reader}
\title{jy_text_file_reader}
\usage{
jy_text_file_reader(file_path)
}
\arguments{
\item{file_path}{the file path to text file to be read (file-format must be "*.txt")}
}
\value{
Vector of char-strings
}
\description{
Reads words from a text file, format them into lower case
}
\examples{
jy_text_file_reader("../data/Example_negative.txt")
}
